+++ src/main/java/com/sensedia/university/functionalities/AlunoFuncionality.java
@@ -1,0 +1,11 @@
+package com.sensedia.university.functionalities;
+
+import com.sensedia.university.models.Aluno;
+
+public interface AlunoFuncionality {
+    void showAlunoOptions();
+    void createAluno(Aluno aluno);
+    void createAlunoCurso();
+    void showAllAlunos();
+    void createAlunoDisciplina();
+}
