+++ src/main/java/com/sensedia/university/functionalities/DisciplinaFuncionality.java
@@ -1,0 +1,10 @@
+package com.sensedia.university.functionalities;
+
+import com.sensedia.university.models.Disciplina;
+
+public interface DisciplinaFuncionality {
+    void showDisciplinaOptions();
+    void createDisciplina(Disciplina disciplina);
+    void updateDisciplina();
+    void showAllDisciplinas();
+}
