+++ src/main/java/com/sensedia/university/functionalities/DocenteFuncionality.java
@@ -1,0 +1,10 @@
+package com.sensedia.university.functionalities;
+
+import com.sensedia.university.models.Docente;
+
+public interface DocenteFuncionality {
+    void showDocenteOptions();
+    void createDocente(Docente docente);
+    void createDocenteDisciplina();
+    void showAllDocentes();
+}
