+++ src/main/java/com/sensedia/university/functionalities/InputFuncionality.java
@@ -1,0 +1,8 @@
+package com.sensedia.university.functionalities;
+
+public interface InputFuncionality {
+    String inputString(Integer limit);
+    String inputMatricula();
+    Integer inputInteger();
+    Integer inputIntegerWithRange(Integer maxRange);
+}
