+++ src/main/java/com/sensedia/university/functionalities/MenuFuncionality.java
@@ -1,0 +1,6 @@
+package com.sensedia.university.functionalities;
+
+public interface MenuFuncionality {
+    void showWelcomeMessage();
+    Integer showMenuOptions();
+}
