+++ src/main/java/com/sensedia/university/repositories/DocenteRepository.java
@@ -1,5 +1,6 @@
 package com.sensedia.university.repositories;
 
+import com.sensedia.university.models.Disciplina;
 import com.sensedia.university.models.Docente;
 
 import java.util.List;
@@ -8,4 +9,6 @@ public interface DocenteRepository {
 
     Docente createDocente(Docente docente);
     List<Docente> getAllDocente();
+    void addDisciplina(Docente docente, Disciplina disciplina);
+    Docente getDocenteByMatricula(String matricula);
 }
