+++ src/main/java/com/sensedia/university/services/DocenteService.java
@@ -1,5 +1,6 @@
 package com.sensedia.university.services;
 
+import com.sensedia.university.models.Disciplina;
 import com.sensedia.university.models.Docente;
 
 import java.util.List;
@@ -7,4 +8,6 @@ import java.util.List;
 public interface DocenteService {
     Docente createDocente(Docente docente);
     List<Docente> getAllDocente();
+    void addDisciplina(Docente docente, Disciplina disciplina);
+    Docente getDocenteByMatricula(String matricula);
 }
